fix bug b.cond

+++ src/sim_utilities.h
@@ -2,7 +2,7 @@
 #define SIM_UTILITIES_H
 
 void print_binary(uint32_t number);
-int32_t sign_extend(uint32_t number, int bits);
+int64_t sign_extend(uint32_t number, int bits);
 uint32_t get_bits(uint32_t number, int start, int end);
 uint32_t negate_number(uint32_t number);
